checkpoint: renames "Config/utilities/StaticConfig.ts" to "Config/Static.ts"
