+++ src/features/TextToImage/Config/Static/index.ts
